requirements.txt
# requirements.txt
web3==6.10.0
requests
matplotlib
streamlit
python-dotenv
qrcode
pillow